requirements.txt
#ASR+LLM依赖

fastapi
uvicorn
python-multipart
numpy>=2
funasr>=0.9.6
websockets
modelscope
onnxruntime
datasets
peft
py7zr
rich

#屏幕点击工具依赖
openai
pillow
pyautogui

#RAG依赖
sentence-transformers
watchdog

#打包exe
PyQt5
requests
pyinstaller


#MemOS记忆系统依赖
streamlit
aiohttp
pydantic>=2.0.0

#langchain 依赖
langchain==1.2.7
langchain-deepseek==1.0.1
langchain-community==0.4.1
wikipedia==1.4.0